tests/wpt/tests/css/css-align/parsing/align-items-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing align-items with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-align-items">
<meta name="assert" content="align-items supports the full grammar 'normal | stretch | <baseline-position> | [ <overflow-position>? <self-position> ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("align-items", "normal");
test_valid_value("align-items", "stretch");

// <baseline-position> = [ first | last ]? baseline
test_valid_value("align-items", "baseline");
test_valid_value("align-items", "first baseline", "baseline");
test_valid_value("align-items", "last baseline");

// <overflow-position>? <self-position>
// <overflow-position> = unsafe | safe
// <self-position> = center | start | end | self-start | self-end | flex-start | flex-end
test_valid_value("align-items", "center");
test_valid_value("align-items", "start");
test_valid_value("align-items", "end");
test_valid_value("align-items", "self-start");
test_valid_value("align-items", "self-end");
test_valid_value("align-items", "flex-start");
test_valid_value("align-items", "flex-end");
test_valid_value("align-items", "unsafe center");
test_valid_value("align-items", "safe self-end");
</script>
</body>
</html>
